+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/DisplayItemBrandsController.java
@@ -3,7 +3,7 @@ package com.dressca.web.consumer.controller;
 import java.util.List;
 import java.util.stream.Collectors;
 import com.dressca.applicationcore.applicationservice.DisplayItemApplicationService;
-import com.dressca.web.consumer.controller.dto.display.GetDisplayItemBrandsResponse;
+import com.dressca.web.consumer.controller.dto.displayitem.GetDisplayItemBrandsResponse;
 import com.dressca.web.consumer.mapper.DisplayItemBrandMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
