+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/dto/displayitem/GetDisplayItemBrandsResponse.java
@@ -1,4 +1,4 @@
-package com.dressca.web.consumer.controller.dto.display;
+package com.dressca.web.consumer.controller.dto.displayitem;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
@@ -7,12 +7,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * 陳列品カテゴリの情報を取得する際に用いる dto クラスです。
+ * 陳列品ブランドの情報を取得する際に用いる dto クラスです。
  */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class GetDisplayItemCategoriesResponse {
+public class GetDisplayItemBrandsResponse {
   @NotNull
   private UUID id;
   @NotNull
