+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/dto/displayitem/GetDisplayItemCategoriesResponse.java
@@ -1,24 +1,20 @@
-package com.dressca.web.consumer.controller.dto.display;
+package com.dressca.web.consumer.controller.dto.displayitem;
 
 import jakarta.validation.constraints.NotNull;
-import java.util.List;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * 陳列品の概要を取得する際に用いる dto クラスです。
+ * 陳列品カテゴリの情報を取得する際に用いる dto クラスです。
  */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class DisplayItemSummaryApiModel {
+public class GetDisplayItemCategoriesResponse {
   @NotNull
   private UUID id;
   @NotNull
   private String name;
-  @NotNull
-  private String productCode;
-  private List<String> assetCodes;
 }
