+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/dto/displayitem/GetDisplayItemResponse.java
@@ -1,4 +1,4 @@
-package com.dressca.web.consumer.controller.dto.display;
+package com.dressca.web.consumer.controller.dto.displayitem;
 
 import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
