+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/dto/displayitem/PagedListOfGetDisplayItemResponse.java
@@ -1,4 +1,4 @@
-package com.dressca.web.consumer.controller.dto.display;
+package com.dressca.web.consumer.controller.dto.displayitem;
 
 import java.util.List;
 import lombok.Data;
