+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/controller/dto/order/OrderItemApiModel.java
@@ -1,9 +1,9 @@
 package com.dressca.web.consumer.controller.dto.order;
 
-import com.dressca.web.consumer.controller.dto.display.DisplayItemSummaryApiModel;
 import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.util.UUID;
+import com.dressca.web.consumer.controller.dto.displayitem.DisplayItemSummaryApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
