+++ samples/web-csr/dressca-backend/web-consumer/src/main/java/com/dressca/web/consumer/mapper/DisplayItemSummaryMapper.java
@@ -2,7 +2,7 @@ package com.dressca.web.consumer.mapper;
 
 import com.dressca.applicationcore.displayitem.DisplayItem;
 import com.dressca.applicationcore.displayitem.DisplayItemAsset;
-import com.dressca.web.consumer.controller.dto.display.DisplayItemSummaryApiModel;
+import com.dressca.web.consumer.controller.dto.displayitem.DisplayItemSummaryApiModel;
 import java.util.List;
 import java.util.stream.Collectors;
